+++ include/Cue.h
@@ -21,10 +21,10 @@ namespace cinder
 		class Cue : public Sequenceable
   		{
 		  public:
-			Cue( std::function<void ()> action, double atTime=0.0 );
+			Cue( std::function<void ()> action, float atTime=0.0 );
 			~Cue();
 			void addReceiver( std::function<void ()> action );
-			virtual void update( double relativeTime );
+			virtual void update( float relativeTime );
 			
 			virtual void complete() {} // completion is a no-op for Cues
