+++ include/Sequenceable.hpp
@@ -18,26 +18,26 @@ namespace cinder
 		class Sequenceable
 		{
 		public:
-			Sequenceable( void *target, double startTime, double duration );
+			Sequenceable( void *target, float startTime, float duration );
 			Sequenceable() : mTarget( 0 ), mDuration( 0 ), mAutoRemove( false ) {}
 			virtual ~Sequenceable() {}
 			
 			void *getTarget() const { return mTarget; }
 			
 			//! go to a specific time
-			void stepTo( double time );
-			virtual void update( double relativeTime ) = 0;
+			void stepTo( float time );
+			virtual void update( float relativeTime ) = 0;
 			virtual void complete() = 0;
 			
-			double getStartTime() const { return mStartTime; }
-			double getEndTime() const { return mStartTime + mDuration; }
-			double getDuration() const { return mDuration; }			
+			float getStartTime() const { return mStartTime; }
+			float getEndTime() const { return mStartTime + mDuration; }
+			float getDuration() const { return mDuration; }			
 			
 			//! set the action's start time
-			void setStartTime( double time ) { mStartTime = time; }
-			void setDuration( double duration ) { mDuration = duration; }
+			void setStartTime( float time ) { mStartTime = time; }
+			void setDuration( float duration ) { mDuration = duration; }
 			//! push back the action's start time
-			void delay( double amt ) { mStartTime += amt; }
+			void delay( float amt ) { mStartTime += amt; }
 			
 			//! change how time behaves
 			void reverse( bool isReversed = true ){ mReversed = isReversed; }
@@ -54,7 +54,7 @@ namespace cinder
 			
 		  protected:
 			void	*mTarget;
-			double	mStartTime, mDuration;
+			float	mStartTime, mDuration;
 			bool	mComplete;
 			bool	mReversed;
 			bool	mAutoRemove;
