+++ include/Tween.cpp
@@ -14,7 +14,7 @@
 
 namespace cinder { namespace tween {
 
-TweenBase::TweenBase( void *target, double startTime, double duration, EaseFn easeFunction )
+TweenBase::TweenBase( void *target, float startTime, float duration, EaseFn easeFunction )
 	: Sequenceable( target, startTime, duration ), mEaseFunction( easeFunction )
 {
 }
